+++ lib/glimmer/dsl/web/element_expression.rb
@@ -21,7 +21,7 @@ module Glimmer
         def add_content(parent, keyword, *args, &block)
           if parent.rendered? || parent.skip_content_on_render_blocks?
             return_value = super(parent, keyword, *args, &block)
-            if return_value.is_a?(String)
+            if return_value.is_a?(String) && parent.dom_element.text.to_s.empty?
               parent.add_text_content(return_value)
             end
             parent.post_add_content